+++ cmd/tenv/tenv.go
@@ -27,6 +27,7 @@ import (
 	"github.com/tofuutils/tenv/v2/config"
 	"github.com/tofuutils/tenv/v2/versionmanager"
 	"github.com/tofuutils/tenv/v2/versionmanager/builder"
+	"github.com/tofuutils/tenv/v2/versionmanager/proxy"
 	terragruntparser "github.com/tofuutils/tenv/v2/versionmanager/semantic/parser/terragrunt"
 
 	"github.com/spf13/cobra"
@@ -64,13 +65,23 @@ func main() {
 		os.Exit(1)
 	}
 
-	if err = initRootCmd(&conf).Execute(); err != nil {
+	builders := map[string]builder.BuilderFunc{
+		config.TofuName:       builder.BuildTofuManager,
+		config.TerraformName:  builder.BuildTfManager,
+		config.TerragruntName: builder.BuildTgManager,
+		config.AtmosName:      builder.BuildAtmosManager,
+	}
+
+	gruntParser := terragruntparser.Make()
+	manageHiddenCallCmd(&conf, builders, gruntParser) // proxy call use os.Exit when called
+
+	if err = initRootCmd(&conf, builders, gruntParser).Execute(); err != nil {
 		fmt.Println(err) //nolint
 		os.Exit(1)
 	}
 }
 
-func initRootCmd(conf *config.Config) *cobra.Command {
+func initRootCmd(conf *config.Config, builders map[string]builder.BuilderFunc, gruntParser terragruntparser.TerragruntParser) *cobra.Command {
 	rootCmd := &cobra.Command{
 		Use:     config.TenvName,
 		Long:    "tenv help manage several versions of OpenTofu (https://opentofu.org), Terraform (https://www.terraform.io), Terragrunt (https://terragrunt.gruntwork.io), and Atmos (https://atmos.tools/).",
@@ -90,8 +101,7 @@ func initRootCmd(conf *config.Config) *cobra.Command {
 		needToken:  true, remoteEnvName: config.TofuRemoteURLEnvName,
 		pRemote: &conf.Tofu.RemoteURL, pPublicKeyPath: &conf.TofuKeyPath,
 	}
-	gruntParser := terragruntparser.Make()
-	tofuManager := builder.BuildTofuManager(conf, gruntParser)
+	tofuManager := builders[config.TofuName](conf, gruntParser)
 	initSubCmds(rootCmd, conf, tofuManager, tofuParams) // add tofu management at root level
 
 	tofuCmd := &cobra.Command{
@@ -116,7 +126,7 @@ func initRootCmd(conf *config.Config) *cobra.Command {
 		needToken: false, remoteEnvName: config.TfRemoteURLEnvName,
 		pRemote: &conf.Tf.RemoteURL, pPublicKeyPath: &conf.TfKeyPath,
 	}
-	initSubCmds(tfCmd, conf, builder.BuildTfManager(conf, gruntParser), tfParams)
+	initSubCmds(tfCmd, conf, builders[config.TerraformName](conf, gruntParser), tfParams)
 
 	rootCmd.AddCommand(tfCmd)
 
@@ -130,7 +140,7 @@ func initRootCmd(conf *config.Config) *cobra.Command {
 	tgParams := subCmdParams{
 		needToken: true, remoteEnvName: config.TgRemoteURLEnvName, pRemote: &conf.Tg.RemoteURL,
 	}
-	initSubCmds(tgCmd, conf, builder.BuildTgManager(conf, gruntParser), tgParams)
+	initSubCmds(tgCmd, conf, builders[config.TerragruntName](conf, gruntParser), tgParams)
 
 	rootCmd.AddCommand(tgCmd)
 
@@ -143,11 +153,24 @@ func initRootCmd(conf *config.Config) *cobra.Command {
 	atmosParams := subCmdParams{
 		needToken: true, remoteEnvName: config.AtmosRemoteURLEnvName, pRemote: &conf.Atmos.RemoteURL,
 	}
-	initSubCmds(atmosCmd, conf, builder.BuildAtmosManager(conf, gruntParser), atmosParams)
+	initSubCmds(atmosCmd, conf, builders[config.AtmosName](conf, gruntParser), atmosParams)
 
 	rootCmd.AddCommand(atmosCmd)
 
 	return rootCmd
+}
+
+func manageHiddenCallCmd(conf *config.Config, builders map[string]builder.BuilderFunc, gruntParser terragruntparser.TerragruntParser) {
+	if len(os.Args) < 3 || os.Args[1] != config.CallSubCmd {
+		return
+	}
+
+	calledNamed, cmdArgs := os.Args[2], os.Args[3:]
+	if builder, ok := builders[calledNamed]; ok {
+		proxy.Exec(conf, builder, gruntParser, calledNamed, cmdArgs)
+	} else if calledNamed == config.AgnosticCmdName {
+		proxy.ExecAgnostic(conf, builders, gruntParser, cmdArgs)
+	}
 }
 
 func newVersionCmd() *cobra.Command {
