+++ cmd/tf/tf.go
@@ -19,58 +19,10 @@
 package main
 
 import (
-	"fmt"
-	"os"
-
 	"github.com/tofuutils/tenv/v2/config"
-	"github.com/tofuutils/tenv/v2/versionmanager/builder"
-	"github.com/tofuutils/tenv/v2/versionmanager/proxy"
-	terragruntparser "github.com/tofuutils/tenv/v2/versionmanager/semantic/parser/terragrunt"
+	lightproxy "github.com/tofuutils/tenv/v2/versionmanager/proxy/light"
 )
 
 func main() {
-	conf, err := config.InitConfigFromEnv()
-	if err != nil {
-		fmt.Println("Configuration error :", err) //nolint
-		os.Exit(1)
-	}
-
-	conf.InitDisplayer(true)
-	gruntParser := terragruntparser.Make()
-	manager := builder.BuildTofuManager(&conf, gruntParser)
-	detectedVersion, err := manager.ResolveWithVersionFiles()
-	if err != nil {
-		fmt.Println("Failed to resolve a version allowing to call tofu :", err) //nolint
-		os.Exit(1)
-	}
-
-	execName := config.TofuName
-	if detectedVersion == "" {
-		execName = config.TerraformName
-		manager = builder.BuildTfManager(&conf, gruntParser)
-		detectedVersion, err = manager.ResolveWithVersionFiles()
-		if err != nil {
-			fmt.Println("Failed to resolve a version allowing to call terraform :", err) //nolint
-			os.Exit(1)
-		}
-
-		if detectedVersion == "" {
-			fmt.Println("No version files found corresponding to opentofu or terraform") //nolint
-			os.Exit(1)
-		}
-	}
-
-	installPath, err := manager.InstallPath()
-	if err != nil {
-		fmt.Println("Failed to create installation directory for", execName, ":", err) //nolint
-		os.Exit(1)
-	}
-
-	detectedVersion, err = manager.Evaluate(detectedVersion, true)
-	if err != nil {
-		fmt.Println("Failed to evaluate the requested version in a specific version allowing to call", execName, ":", err) //nolint
-		os.Exit(1)
-	}
-
-	proxy.RunCmd(installPath, detectedVersion, execName)
+	lightproxy.Exec(config.AgnosticCmdName)
 }
